Sulong: Fix truffle-c test warnings

+++ sulong/tests/com.oracle.truffle.llvm.tests.sulong.native/c/truffle-c/unsignedLongTest/divTest3.c
@@ -30,5 +30,5 @@
 int main() {
     unsigned long a = -1;
     unsigned long b = 1;
-    return (a / b) == -1;
+    return (int) (a / b) == -1;
 }
